directory: move playlist components to playlist folder
